Stop hard-coding float32 in PMCMC/ABC dtype paths
Several arrays in proposals.py and Pomp.pmcmc/abc were constructed with
an explicit dtype=jnp.float32, which broke scan-carry type checking
when jax_enable_x64 is True (the carry's input theta is float32 but
the body returns float64 from jax.random.normal etc., and scan
rejects the mismatch).

Drop the explicit dtypes and let jax's configured default precision
decide.  Adaptive proposal state and abc's epsilon now follow
theta_arr.dtype so they're consistent under both x32 and x64.

Verified with a PMCMC run on the Chopin Ch.14 thalamus model
(x64 enabled) and the pmcmc/abc test suites (47 passing).

+++ pypomp/core/algorithms/abc.py
@@ -74,7 +74,7 @@ def _abc_internal(
     * ``theta_trace``: shape ``(Nabc + 1, d)``.
     * ``accepts``: scalar count of accepted proposals.
     """
-    eps2 = jnp.asarray(epsilon, dtype=jnp.float32) ** 2
+    eps2 = jnp.asarray(epsilon, dtype=theta_arr.dtype) ** 2
 
     def _sim_distance(theta: jax.Array, sim_key: jax.Array) -> jax.Array:
         _, Y = _simulate_internal(
